backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp>=2.10.6",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
